requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
torchvision
numpy
packaging
tqdm
ftfy
regex